before: widen the stamp split scanner's depth counter to i64
parse_clock_str counted paren depth in an i32, so a ~2 GiB run of '('
overflowed it: a panic on public text input in overflow-checked builds,
a wrapped depth (wrong error) in release. An i64 depth preserves the
scanner's exact split semantics — negative depths from unmatched ')'
still participate in the depth-0 comma test — and cannot overflow, since
depth moves at most one per input byte and an allocation holds at most
isize::MAX (< 2^63) bytes.

Pinned by a proptest family over deep-nesting stamp texts (balanced,
over- and under-closed, comma at any depth) asserting Clock::from_str
never panics, plus an #[ignore]d ~2 GiB deterministic witness feeding
2^31 + 1 opens through Clock::from_str and asserting graceful rejection
(red before the widening: 'attempt to add with overflow'; green after).

+++ crates/before/src/codec/text.rs
@@ -189,7 +189,9 @@ pub(crate) fn parse_clock_str(s: &str) -> Result<(BitsMut, &str), Parse> {
         return Err(Parse::Syntax);
     }
     let inner = &t[1..t.len() - 1];
-    let mut depth: i32 = 0;
+    // i64 cannot overflow: depth moves by at most one per input byte, and an
+    // allocation holds at most `isize::MAX` (< 2⁶³) bytes.
+    let mut depth: i64 = 0;
     let mut split = None;
     for (k, &c) in inner.as_bytes().iter().enumerate() {
         match c {
